# swagger-spec.yaml
openapi: 3.0.0
info:
  title: API de Postagem Social
  version: 1.0.0
  description: API para postar conteúdo em diversas redes sociais.
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SocialPostRequest:
      type: object
      properties:
        text:
          type: string
        images:
          type: array
          items:
            type: string
            format: byte
        tags:
          type: array
          items:
            type: string
paths:
...
  /twitter/post:
    post:
      summary: Cria um novo tweet.
      tags:
        - Twitter
      description: >
        Publica um novo tweet na conta do Twitter autenticada. A postagem pode
        conter texto, até 4 imagens e tags.


        **Corpo da Requisição:**

        * **`text`** (string, obrigatório): O conteúdo principal do tweet.

        * **`images`** (array, opcional): Uma lista de até 4 imagens no formato Data URL (base64).

        * **`tags`** (array, opcional): Uma lista de tags que serão convertidas em hashtags e adicionadas ao final do texto.


        **Corpo da Resposta:**

        * Retorna um objeto com os dados do tweet criado com sucesso (status `201 Created`).

        * Retorna um erro `400 Bad Request` se os dados forem inválidos (ex: mais de 4 imagens).
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/SocialPostRequest"
                - type: object
                  properties:
                    text:
                      type: string
                    images:
                      type: array
                      items:
                        type: base64
                    tags:
                      type: array
                      items:
                        type: string
            example:
              text: Este é um tweet de exemplo!
              tags:
                - api
                - teste
              images:
                - data:image/png;base64,iVBORw0KGgo...
      responses:
        "201":
          description: Tweet criado com sucesso.
...